requirements.txt
Flask==3.1.1
Werkzeug==3.1.3
gunicorn==21.2.0
Flask-Login==0.6.3
google-cloud-storage==3.1.0 